requirements.txt
beautifulsoup4==4.12.3
langchain_community==0.3.16
langchain_openai==0.3.3
pinecone==5.4.2
python-dotenv==1.0.1
Requests==2.32.3
